// src/core/plugins/retro-bus/shared/enums.ts
export enum RETRO_BUS_METAS {
    EXTRA_STATE = 'retro-bus:extraState',
    AMOUNT_ATTACHED = 'retro-bus:amountAttached',
    ATTACHED_VEHICLES = 'retro-bus:attachedVehicles',
}

export enum RETRO_BUS_EVENTS {
    ADD_BIKE_MENU = 'retro-bus:attachBikeMenu',
    REMOVE_BIKE_MENU = 'retro-bus:detachBikeMenu',
    ATTACH_BIKE = 'retro-bus:attachBike',
    DETACH_BIKE = 'retro-bus:detachBike',
    BIKE_LOCK = 'retro-bus:toggleBikeLock',
}